requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.4
Flask==1.1.1
functions-framework==1.2.0
idna==2.9
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pathtools==0.1.2
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
watchdog==0.10.2
Werkzeug==1.0.0
